#4 base class for custom messages
