+++ src/Subscriber/Channel/SimpleArrayChannelMessage.php
@@ -4,13 +4,10 @@ declare(strict_types=1);
 
 namespace App\Subscriber\Channel;
 
-use App\Entity\PaymentProcessing;
 use App\Subscriber\Exception\ChannelMessageException;
 
-final class SimpleArrayChannelMessage implements ChannelMessageInterface
+final class SimpleArrayChannelMessage extends AbstractChannelMessage
 {
-    private PaymentProcessing | null $paymentProcessing = null;
-
     /**
      * @return array{
      *     order_num: string,
@@ -33,16 +30,5 @@ final class SimpleArrayChannelMessage implements ChannelMessageInterface
             //'transaction_id' => $this->getResponse()->getTransactionId(),
             'response' => $paymentProcessing->getResponseData(),
         ];
-    }
-
-    public function setPaymentProcessing(PaymentProcessing $paymentProcessing): void
-    {
-        $this->paymentProcessing = $paymentProcessing;
-    }
-
-    public function getPaymentProcessing(): PaymentProcessing
-    {
-        return $this->paymentProcessing
-            ?? throw new ChannelMessageException('Payment processing response not set.');
     }
 }
